refactor: rename file JsonConfigLoader
